Introduce uploadWorker
