--- packages/test-utils/src/legacy-cli/index.ts
@@ -1,1 +1,0 @@
-export { listAllRecordings, removeRecording, uploadRecording } from "./main";
