--- packages/test-utils/src/legacy-cli/updateStatus.ts
@@ -1,14 +1,0 @@
-import { RecordingEntry } from "./types";
-
-export function updateStatus(recording: RecordingEntry, status: RecordingEntry["status"]) {
-  // Once a recording enters an unusable or crashed status, don't change it
-  // except to mark crashes as uploaded.
-  if (
-    recording.status == "unusable" ||
-    recording.status == "crashUploaded" ||
-    (recording.status == "crashed" && status != "crashUploaded")
-  ) {
-    return;
-  }
-  recording.status = status;
-}
